refactor: request header generation
* should work in NodeJS now
* DRYed up code
