+++ src/call/index.ts
@@ -1,17 +1,18 @@
 import axios, { AxiosRequestHeaders } from 'axios';
 import { Error } from '../frappe_app/types';
+import { getRequestHeaders } from '../utils';
 export class FrappeCall {
   /** URL of the Frappe App instance */
   private readonly appURL: string;
 
-  /** Whether to use the token from the window object */
+  /** Whether to use the token based auth */
   readonly useToken: boolean;
 
   /** Token to be used for authentication */
   readonly token?: () => string;
 
   /** Type of token to be used for authentication */
-  readonly tokenType?: 'Bearer' | 'token'
+  readonly tokenType?: 'Bearer' | 'token';
 
   constructor(appURL: string, useToken?: boolean, token?: () => string, tokenType?: 'Bearer' | 'token') {
     this.appURL = appURL;
@@ -20,21 +21,13 @@ export class FrappeCall {
     this.tokenType = tokenType;
   }
 
+  private getPreparedRequestHeaders(): AxiosRequestHeaders {
+    return getRequestHeaders(this.useToken, this.tokenType, this.token);
+  }
+
   /** Makes a GET request to the specified endpoint */
   async get<T = any>(path: string, params?: Record<string, any>): Promise<T> {
-    const headers: AxiosRequestHeaders = {
-      Accept: 'application/json',
-      'Content-Type': 'application/json; charset=utf-8',
-      'X-Frappe-Site-Name': window.location.hostname,
-    };
-
-    if ((window as any).csrf_token && (window as any).csrf_token !== '{{ csrf_token }}') {
-      headers['X-Frappe-CSRF-Token'] = (window as any).csrf_token;
-    }
-
-    if (this.useToken && this.tokenType && this.token) {
-      headers.Authorization = `${this.tokenType} ${this.token()}`;
-    }
+    const headers = this.getPreparedRequestHeaders();
 
     return axios
       .get(`${this.appURL}/api/method/${path}`, {
@@ -55,19 +48,7 @@ export class FrappeCall {
 
   /** Makes a POST request to the specified endpoint */
   async post<T = any>(path: string, params?: any): Promise<T> {
-    const headers: AxiosRequestHeaders = {
-      Accept: 'application/json',
-      'Content-Type': 'application/json; charset=utf-8',
-      'X-Frappe-Site-Name': window.location.hostname,
-    };
-
-    if ((window as any).csrf_token && (window as any).csrf_token !== '{{ csrf_token }}') {
-      headers['X-Frappe-CSRF-Token'] = (window as any).csrf_token;
-    }
-
-    if (this.useToken && this.tokenType && this.token) {
-      headers.Authorization = `${this.tokenType} ${this.token()}`;
-    }
+    const headers = this.getPreparedRequestHeaders();
 
     return axios
       .post(
@@ -93,19 +74,7 @@ export class FrappeCall {
 
   /** Makes a PUT request to the specified endpoint */
   async put<T = any>(path: string, params?: any): Promise<T> {
-    const headers: AxiosRequestHeaders = {
-      Accept: 'application/json',
-      'Content-Type': 'application/json; charset=utf-8',
-      'X-Frappe-Site-Name': window.location.hostname,
-    };
-
-    if ((window as any).csrf_token && (window as any).csrf_token !== '{{ csrf_token }}') {
-      headers['X-Frappe-CSRF-Token'] = (window as any).csrf_token;
-    }
-
-    if (this.useToken && this.tokenType && this.token) {
-      headers.Authorization = `${this.tokenType} ${this.token()}`;
-    }
+    const headers = this.getPreparedRequestHeaders();
 
     return axios
       .put(
@@ -131,19 +100,7 @@ export class FrappeCall {
 
   /** Makes a DELETE request to the specified endpoint */
   async delete<T = any>(path: string, params?: any): Promise<T> {
-    const headers: AxiosRequestHeaders = {
-      Accept: 'application/json',
-      'Content-Type': 'application/json; charset=utf-8',
-      'X-Frappe-Site-Name': window.location.hostname,
-    };
-
-    if ((window as any).csrf_token && (window as any).csrf_token !== '{{ csrf_token }}') {
-      headers['X-Frappe-CSRF-Token'] = (window as any).csrf_token;
-    }
-
-    if (this.useToken && this.tokenType && this.token) {
-      headers.Authorization = `${this.tokenType} ${this.token()}`;
-    }
+    const headers = this.getPreparedRequestHeaders();
 
     return axios
       .delete(`${this.appURL}/api/method/${path}`, {
